Make Conscrypt initialization explicit
